Finished Project 1B

+++ proj1b/OffByN.java
@@ -7,7 +7,6 @@ public class OffByN implements CharacterComparator {
 
     @Override
     public boolean equalChars(char x, char y) {
-        //return Math.abs(Character.toUpperCase(x) - Character.toUpperCase(y)) == N;
-        return Math.abs((int) x - (int) y) == N;
+        return Math.abs(x - y) == N;
     }
 }
